tests/tokenizer[23]: fix handling of CHARACTER tokens
Where a CHARACTER token is emitted but the expected data is shorter
than the data in the token, we want to consume the expected data
and then process the remaining token data as if it were emitted
separately. Sadly this didn't happen as we never initialised the
replacement token correctly. Make this so, and also tell the
compiler that the recursive call is actually a tail call so it
can optimise it appropriately.